Connect and send data via sockjs to client
