+++ server/views/arViews/d3.ejs
@@ -3,28 +3,25 @@
 
 	<h1>Homepage</h1>
 	<p><%= message %></p>
-  <pre><%= h.dump(data) %></pre>
-
-
-
-  <svg width="960" height="500"></svg>
-
-  <script src="https://d3js.org/d3.v4.min.js"></script>
 
   <script type="text/javascript">
 		var sock = new SockJS('http://localhost:9999/echo');
+		// var sock = new SockJS('https://0214295b.ngrok.io');
 		sock.onopen = function() {
-				console.log('open');
-				sock.send('test');
+			console.log('open');
+			// sock.send('bal');
 		};
 		
 		sock.onmessage = function(e) {
-				console.log('message', e.data);
-				sock.close();
+			// var stomp = JSON.parse(e.data);
+			// console.log(e.data, stomp)
+			console.log('message', e.data);
+			// sock.close();
+			// e.data.init();
 		};
 		
 		sock.onclose = function() {
-				console.log('close');
+			console.log('close');
 		};
   </script>
